Move `napi5` --> `napi`-- round 1 (#39511)
This is one step in the slow strangulation of `napi`. This removes
`napi5`.
- static helper methods from `napi5` are moved to a separate file,
`napi_helper` (which will eventually be deleted)
- methods which are simply called are moved into `napi` directly.

Part of Issue #38332 

#### Further detail of work
Some functions were also deleted along the way, when it was clear they
have no real usage inside Mantid.
